requirements.txt
agent-framework==1.0.0
azure-identity==1.25.1
apify-client==2.5.0
streamlit==1.54.0
pandas==2.3.2
plotly==6.5.2
python-dotenv==1.1.1
pydantic==2.13.0b2
nest-asyncio==1.6.0
streamlit-autorefresh==1.0.1
fpdf2==2.8.7
mailtrap==2.4.0
